requirements.txt
PyMuPdf
ensure 
langchain
dotenv
tiktoken 
python-dotenv
openai
neo4j 
pypdf
streamlit
Pillow
transformers
streamlit-extras
torch
tokenizers